Add properties to `Aspect`. Also change the type signature and yield value of `Aspect.xor`
I needed to be able to easily get the flattened set of all sets in `Aspect.either_or` as
`Aspect.flat_either_or` and to easily get the union of all (`Aspect.all`) the sets--
saves A LOT of typing on my part.

The change to `Aspect.xor` was warranted because it yields sets of length 1 anyway,
so why yield *that* rather than the *one* element in the set?