Add two more properties, component names validity check function, and minor fixes to error messages.
I added the properties `live_components` and `registered_components` because I wanted
to get the set of registered components and live entities (those two are separate properties)--
these are more semantically meaningful as well.

The function to check the validity of components name (`_is_component_names_valid`) does so by checking
if the input is a subset or equal to the registered components (both are sets); this means
that checking for validity is far faster (because it uses set operations) instead of looping
and checking each element every. single. time.
The reason I added this is kinda self-explanatory-- saving time|more meaningful checks.