Add better type annotations and input checking
In some places, `components` was used rather than `new_components`-- which was
the more specific type (`new_components`). I fixed it to used `new_components`
wherever it was needed.

I also made `create_entity` catch when the (required) 'args' or 'kwargs' keys
don't exist if they call the method where `instantiated` is True-- it raises
`InvalidEntityIDError` in the handler for `KeyError`.